requirements.txt
aiocron==1.8.*
asyncio==3.4.*
discord.py==2.1.*
mcrcon==0.7.*
pillow==9.3.*
psutil==5.9.*
psycopg2==2.9.*
pyyaml==6.0.*
requests==2.28.*
